+++ convertkit/resources/frontend/js/convertkit.js
@@ -7,60 +7,6 @@
  */
 
 /**
- * Gets the subscriber ID for the given email address, storing
- * it in the `ck_subscriber_id` cookie if it exists.
- *
- * Typically called when the user completes a ConvertKit Form
- * that has either "Auto-confirm new subscribers" or
- * "Send subscriber to thank you page" enabled (both scenarios
- * include a ck_subscriber_id).
- *
- * @since   1.9.6
- *
- * @param {string} emailAddress Email Address
- */
-function convertStoreSubscriberEmailAsIDInCookie(emailAddress) {
-	if (convertkit.debug) {
-		console.log('convertStoreSubscriberEmailAsIDInCookie');
-		console.log(emailAddress);
-	}
-
-	fetch(convertkit.ajaxurl, {
-		method: 'POST',
-		headers: {
-			'Content-Type': 'application/x-www-form-urlencoded',
-			'X-WP-Nonce': convertkit.nonce,
-		},
-		body: new URLSearchParams({
-			email: emailAddress,
-		}),
-	})
-		.then(function (response) {
-			if (convertkit.debug) {
-				console.log(response);
-			}
-
-			return response.json();
-		})
-		.then(function (result) {
-			if (convertkit.debug) {
-				console.log(result);
-			}
-
-			// Emit custom event with subscriber ID.
-			convertKitEmitCustomEvent('convertkit_user_subscribed', {
-				id: result.id,
-				email: emailAddress,
-			});
-		})
-		.catch(function (error) {
-			if (convertkit.debug) {
-				console.error(error);
-			}
-		});
-}
-
-/**
  * Remove the url subscriber_id url param
  *
  * The 'ck_subscriber_id' should only be set on URLs included on
@@ -104,21 +50,6 @@ function convertKitRemoveSubscriberIDFromURL(url) {
 	convertKitEmitCustomEvent('kit_subscriber_id_removed_from_url', {
 		id: ck_subscriber_id,
 	});
-}
-
-/**
- * Utility function to pause for the given number of milliseconds
- *
- * @since   1.9.6
- * @param {number} milliseconds Number of milliseconds to pause for.
- */
-function convertKitSleep(milliseconds) {
-	const start = new Date().getTime();
-	for (let i = 0; i < 1e7; i++) {
-		if (new Date().getTime() - start > milliseconds) {
-			break;
-		}
-	}
 }
 
 /**
@@ -170,51 +101,6 @@ if (typeof convertkit !== 'undefined') {
 	document.addEventListener('DOMContentLoaded', function () {
 		// Removes `ck_subscriber_id` from the URI.
 		convertKitRemoveSubscriberIDFromURL(window.location.href);
-
-		// Store subscriber ID as a cookie from the email address used when a ConvertKit Form is submitted.
-		document.addEventListener('click', function (e) {
-			// Check if the form submit button was clicked, or the span element was clicked and its parent is the form submit button.
-			if (
-				!e.target.matches('.formkit-submit') &&
-				(!e.target.parentElement ||
-					!e.target.parentElement.matches('.formkit-submit'))
-			) {
-				if (convertkit.debug) {
-					console.log('not a ck form');
-				}
-
-				return;
-			}
-
-			// Get email address.
-			const emailAddress = document.querySelector(
-				'input[name="email_address"]'
-			).value;
-
-			// If the email address is empty, don't attempt to get the subscriber ID by email.
-			if (!emailAddress.length) {
-				if (convertkit.debug) {
-					console.log('email empty');
-				}
-
-				return;
-			}
-
-			// If the email address is invalid, don't attempt to get the subscriber ID by email.
-			const validator =
-				/^(([^<>()\[\]\\.,;:\s@"]+(\.[^<>()\[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/;
-			if (!validator.test(emailAddress.toLowerCase())) {
-				if (convertkit.debug) {
-					console.log('email not an email address');
-				}
-
-				return;
-			}
-
-			// Wait a moment before sending the AJAX request.
-			convertKitSleep(2000);
-			convertStoreSubscriberEmailAsIDInCookie(emailAddress);
-		});
 
 		// Set a cookie if any scripts with data-kit-limit-per-session attribute exist.
 		if (